requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.2.6
docker==7.0.0
idna==3.6
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
pytz==2024.1
requests==2.31.0
setuptools==69.2.0
sqlparse==0.4.4
urllib3==2.2.1
wheel==0.43.0
